Client side changes for login module
