+++ ems-client/proxy.conf.json
@@ -2,6 +2,9 @@
     "/server": {
       "target": "http://localhost:8080",
       "secure": false,
-      "changeOrigin": true
+      "logLevel": "debug",
+      "pathRewrite": {
+        "^/server": ""
+      }
     }
   }
