+++ ems-client/src/app/nav/nav.component.html
@@ -1,0 +1,3 @@
+<p>
+  nav works!
+</p>
