--- my-first-project/src/app/app-routing.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-// import { LoginComponent } from './login/login.component';
-
-const routes: Routes = [
-//   {
-//   path: 'login',
-//   component: LoginComponent,
-// },
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
-})
-export class AppRoutingModule { }
