--- my-first-project/src/app/app.component.ts
@@ -1,17 +1,0 @@
-import { Component } from '@angular/core';
-import { ApiService } from './api.service';
-
-
-@Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
-})
-export class AppComponent {
-  title = 'EMS';
-  constructor(public dataService: ApiService) { } 
- myFunct(){
-   window.location.href='http://localhost:8080/'
-  //  this.dataService.Login();
-}
-}
